Fix build after rebase
